FIX: Move front-end ODE equation tests to analyses.rs
